♻️ refactor(quota): share provider quota request helpers
Move provider-specific quota request and parsing logic into shared quota utilities.

This keeps quota management behavior intact while making the same logic reusable by monitoring.

Low risk: existing quota tests and build pass with the refactored request path.
