+++ src/components/quota/quotaConfigs.ts
@@ -7,66 +7,31 @@ import type { ReactNode } from 'react';
 import type { TFunction } from 'i18next';
 import type {
   AntigravityQuotaGroup,
-  AntigravityModelsPayload,
   AntigravityQuotaState,
   AuthFileItem,
   ClaudeExtraUsage,
-  ClaudeProfileResponse,
   ClaudeQuotaState,
   ClaudeQuotaWindow,
-  ClaudeUsagePayload,
   CodexQuotaState,
   CodexQuotaWindow,
-  CodexUsagePayload,
-  GeminiCliCodeAssistPayload,
-  GeminiCliCredits,
-  GeminiCliParsedBucket,
   GeminiCliQuotaBucketState,
   GeminiCliQuotaState,
-  GeminiCliUserTier,
   KimiQuotaRow,
   KimiQuotaState,
 } from '@/types';
-import {
-  apiCallApi,
-  authFilesApi,
-  getApiCallErrorMessage,
-  requestCodexUsageRaw,
-} from '@/services/api';
 import { useQuotaStore } from '@/stores';
 import {
-  ANTIGRAVITY_QUOTA_URLS,
-  ANTIGRAVITY_REQUEST_HEADERS,
-  CLAUDE_PROFILE_URL,
-  CLAUDE_USAGE_URL,
-  CLAUDE_REQUEST_HEADERS,
-  CLAUDE_USAGE_WINDOW_KEYS,
-  GEMINI_CLI_QUOTA_URL,
-  GEMINI_CLI_CODE_ASSIST_URL,
-  GEMINI_CLI_REQUEST_HEADERS,
-  KIMI_USAGE_URL,
-  KIMI_REQUEST_HEADERS,
-  normalizeGeminiCliModelId,
-  normalizeNumberValue,
   normalizePlanType,
-  normalizeQuotaFraction,
-  normalizeStringValue,
-  parseAntigravityPayload,
-  parseClaudeUsagePayload,
-  parseGeminiCliQuotaPayload,
-  parseGeminiCliCodeAssistPayload,
-  parseKimiUsagePayload,
   resolveCodexChatgptAccountId,
   resolveCodexPlanType,
-  resolveGeminiCliProjectId,
   formatQuotaResetTime,
   formatKimiResetHint,
-  buildAntigravityQuotaGroups,
-  buildCodexQuotaWindowInfos,
-  buildGeminiCliQuotaBuckets,
-  buildKimiQuotaRows,
-  createStatusError,
-  getStatusFromError,
+  fetchAntigravityQuota,
+  fetchClaudeQuota,
+  fetchCodexQuota,
+  fetchGeminiCliCodeAssist,
+  fetchGeminiCliQuotaBuckets,
+  fetchKimiQuota,
   isAntigravityFile,
   isClaudeFile,
   isCodexFile,
@@ -75,7 +40,6 @@ import {
   isKimiFile,
   isRuntimeOnlyAuthFile,
 } from '@/utils/quota';
-import { normalizeAuthIndex } from '@/utils/authIndex';
 import type { QuotaRenderHelpers } from './QuotaCard';
 import styles from '@/features/quota/QuotaPage.module.scss';
 
@@ -84,7 +48,6 @@ type QuotaUpdater<T> = T | ((prev: T) => T);
 type QuotaType = 'antigravity' | 'claude' | 'codex' | 'gemini-cli' | 'kimi';
 export type QuotaSortMode = 'default' | 'name-asc' | 'plan-desc' | 'plan-asc';
 
-const DEFAULT_ANTIGRAVITY_PROJECT_ID = 'bamboo-precept-lgxtn';
 const QUOTA_PROGRESS_HIGH_THRESHOLD = 70;
 const QUOTA_PROGRESS_MEDIUM_THRESHOLD = 30;
 const geminiCliSupplementaryRequestIds = new Map<string, number>();
@@ -131,241 +94,6 @@ export interface QuotaConfig<TState, TData> {
   getPlanSortRank?: (file: AuthFileItem, quota: TState | undefined) => number | null;
   renderQuotaItems: (quota: TState, t: TFunction, helpers: QuotaRenderHelpers) => ReactNode;
 }
-
-const resolveAntigravityProjectId = async (file: AuthFileItem): Promise<string> => {
-  try {
-    const text = await authFilesApi.downloadText(file.name);
-    const trimmed = text.trim();
-    if (!trimmed) return DEFAULT_ANTIGRAVITY_PROJECT_ID;
-
-    const parsed = JSON.parse(trimmed) as Record<string, unknown>;
-    const topLevel = normalizeStringValue(parsed.project_id ?? parsed.projectId);
-    if (topLevel) return topLevel;
-
-    const installed =
-      parsed.installed && typeof parsed.installed === 'object' && parsed.installed !== null
-        ? (parsed.installed as Record<string, unknown>)
-        : null;
-    const installedProjectId = installed
-      ? normalizeStringValue(installed.project_id ?? installed.projectId)
-      : null;
-    if (installedProjectId) return installedProjectId;
-
-    const web =
-      parsed.web && typeof parsed.web === 'object' && parsed.web !== null
-        ? (parsed.web as Record<string, unknown>)
-        : null;
-    const webProjectId = web ? normalizeStringValue(web.project_id ?? web.projectId) : null;
-    if (webProjectId) return webProjectId;
-  } catch {
-    return DEFAULT_ANTIGRAVITY_PROJECT_ID;
-  }
-
-  return DEFAULT_ANTIGRAVITY_PROJECT_ID;
-};
-
-const fetchAntigravityQuota = async (
-  file: AuthFileItem,
-  t: TFunction
-): Promise<AntigravityQuotaGroup[]> => {
-  const rawAuthIndex = file['auth_index'] ?? file.authIndex;
-  const authIndex = normalizeAuthIndex(rawAuthIndex);
-  if (!authIndex) {
-    throw new Error(t('antigravity_quota.missing_auth_index'));
-  }
-
-  const projectId = await resolveAntigravityProjectId(file);
-  const requestBody = JSON.stringify({ project: projectId });
-
-  let lastError = '';
-  let lastStatus: number | undefined;
-  let priorityStatus: number | undefined;
-  let hadSuccess = false;
-
-  for (const url of ANTIGRAVITY_QUOTA_URLS) {
-    try {
-      const result = await apiCallApi.request({
-        authIndex,
-        method: 'POST',
-        url,
-        header: { ...ANTIGRAVITY_REQUEST_HEADERS },
-        data: requestBody,
-      });
-
-      if (result.statusCode < 200 || result.statusCode >= 300) {
-        lastError = getApiCallErrorMessage(result);
-        lastStatus = result.statusCode;
-        if (result.statusCode === 403 || result.statusCode === 404) {
-          priorityStatus ??= result.statusCode;
-        }
-        continue;
-      }
-
-      hadSuccess = true;
-      const payload = parseAntigravityPayload(result.body ?? result.bodyText);
-      const models = payload?.models;
-      if (!models || typeof models !== 'object' || Array.isArray(models)) {
-        lastError = t('antigravity_quota.empty_models');
-        continue;
-      }
-
-      const groups = buildAntigravityQuotaGroups(models as AntigravityModelsPayload);
-      if (groups.length === 0) {
-        lastError = t('antigravity_quota.empty_models');
-        continue;
-      }
-
-      return groups;
-    } catch (err: unknown) {
-      lastError = err instanceof Error ? err.message : t('common.unknown_error');
-      const status = getStatusFromError(err);
-      if (status) {
-        lastStatus = status;
-        if (status === 403 || status === 404) {
-          priorityStatus ??= status;
-        }
-      }
-    }
-  }
-
-  if (hadSuccess) {
-    return [];
-  }
-
-  throw createStatusError(lastError || t('common.unknown_error'), priorityStatus ?? lastStatus);
-};
-
-const buildCodexQuotaWindows = (payload: CodexUsagePayload, t: TFunction): CodexQuotaWindow[] =>
-  buildCodexQuotaWindowInfos(payload).map((window) => ({
-    id: window.id,
-    label: t(window.labelKey, window.labelParams),
-    labelKey: window.labelKey,
-    labelParams: window.labelParams,
-    usedPercent: window.usedPercent,
-    resetLabel: window.resetLabel,
-  }));
-
-const fetchCodexQuota = async (
-  file: AuthFileItem,
-  t: TFunction
-): Promise<{ planType: string | null; windows: CodexQuotaWindow[] }> => {
-  const rawAuthIndex = file['auth_index'] ?? file.authIndex;
-  const authIndex = normalizeAuthIndex(rawAuthIndex);
-  if (!authIndex) {
-    throw new Error(t('codex_quota.missing_auth_index'));
-  }
-
-  const planTypeFromFile = resolveCodexPlanType(file);
-  const accountId = resolveCodexChatgptAccountId(file);
-  const { result, payload } = await requestCodexUsageRaw({ authIndex, accountId });
-
-  if (result.statusCode < 200 || result.statusCode >= 300) {
-    throw createStatusError(getApiCallErrorMessage(result), result.statusCode);
-  }
-
-  if (!payload) {
-    throw new Error(t('codex_quota.empty_windows'));
-  }
-
-  const planTypeFromUsage = normalizePlanType(payload.plan_type ?? payload.planType);
-  const windows = buildCodexQuotaWindows(payload, t);
-  return { planType: planTypeFromUsage ?? planTypeFromFile, windows };
-};
-
-const GEMINI_CLI_G1_CREDIT_TYPE = 'GOOGLE_ONE_AI';
-
-const GEMINI_CLI_TIER_LABELS: Record<string, string> = {
-  'free-tier': 'tier_free',
-  'legacy-tier': 'tier_legacy',
-  'standard-tier': 'tier_standard',
-  'g1-pro-tier': 'tier_pro',
-  'g1-ultra-tier': 'tier_ultra',
-};
-
-const resolveGeminiCliTierLabel = (
-  payload: GeminiCliCodeAssistPayload | null,
-  t: TFunction
-): string | null => {
-  if (!payload) return null;
-  const currentTier: GeminiCliUserTier | null | undefined =
-    payload.currentTier ?? payload.current_tier;
-  const paidTier: GeminiCliUserTier | null | undefined = payload.paidTier ?? payload.paid_tier;
-  const rawId = normalizeStringValue(paidTier?.id) ?? normalizeStringValue(currentTier?.id);
-  if (!rawId) return null;
-  const tierId = rawId.toLowerCase();
-  const labelKey = GEMINI_CLI_TIER_LABELS[tierId];
-  return labelKey ? t(`gemini_cli_quota.${labelKey}`) : rawId;
-};
-
-const resolveGeminiCliTierId = (payload: GeminiCliCodeAssistPayload | null): string | null => {
-  if (!payload) return null;
-  const currentTier: GeminiCliUserTier | null | undefined =
-    payload.currentTier ?? payload.current_tier;
-  const paidTier: GeminiCliUserTier | null | undefined = payload.paidTier ?? payload.paid_tier;
-  const rawId = normalizeStringValue(paidTier?.id) ?? normalizeStringValue(currentTier?.id);
-  return rawId ? rawId.toLowerCase() : null;
-};
-
-const resolveGeminiCliCreditBalance = (
-  payload: GeminiCliCodeAssistPayload | null
-): number | null => {
-  if (!payload) return null;
-  const paidTier: GeminiCliUserTier | null | undefined = payload.paidTier ?? payload.paid_tier;
-  const currentTier: GeminiCliUserTier | null | undefined =
-    payload.currentTier ?? payload.current_tier;
-  const tier = paidTier ?? currentTier;
-  if (!tier) return null;
-  const credits: GeminiCliCredits[] = tier.availableCredits ?? tier.available_credits ?? [];
-  let total = 0;
-  let found = false;
-  for (const credit of credits) {
-    const creditType = normalizeStringValue(credit.creditType ?? credit.credit_type);
-    if (creditType !== GEMINI_CLI_G1_CREDIT_TYPE) continue;
-    const amount = normalizeNumberValue(credit.creditAmount ?? credit.credit_amount);
-    if (amount !== null) {
-      total += amount;
-      found = true;
-    }
-  }
-  return found ? total : null;
-};
-
-const fetchGeminiCliCodeAssist = async (
-  authIndex: string,
-  projectId: string,
-  t: TFunction
-): Promise<{ tierLabel: string | null; tierId: string | null; creditBalance: number | null }> => {
-  try {
-    const result = await apiCallApi.request({
-      authIndex,
-      method: 'POST',
-      url: GEMINI_CLI_CODE_ASSIST_URL,
-      header: { ...GEMINI_CLI_REQUEST_HEADERS },
-      data: JSON.stringify({
-        cloudaicompanionProject: projectId,
-        metadata: {
-          ideType: 'IDE_UNSPECIFIED',
-          platform: 'PLATFORM_UNSPECIFIED',
-          pluginType: 'GEMINI',
-          duetProject: projectId,
-        },
-      }),
-    });
-
-    if (result.statusCode < 200 || result.statusCode >= 300) {
-      return { tierLabel: null, tierId: null, creditBalance: null };
-    }
-
-    const payload = parseGeminiCliCodeAssistPayload(result.body ?? result.bodyText);
-    return {
-      tierLabel: resolveGeminiCliTierLabel(payload, t),
-      tierId: resolveGeminiCliTierId(payload),
-      creditBalance: resolveGeminiCliCreditBalance(payload),
-    };
-  } catch {
-    return { tierLabel: null, tierId: null, creditBalance: null };
-  }
-};
 
 const readGeminiCliSupplementarySnapshot = (
   fileName: string,
@@ -441,65 +169,11 @@ const fetchGeminiCliQuota = async (
   tierId: string | null;
   creditBalance: number | null;
 }> => {
-  const rawAuthIndex = file['auth_index'] ?? file.authIndex;
-  const authIndex = normalizeAuthIndex(rawAuthIndex);
-  if (!authIndex) {
-    throw new Error(t('gemini_cli_quota.missing_auth_index'));
-  }
-
-  const projectId = resolveGeminiCliProjectId(file);
-  if (!projectId) {
-    throw new Error(t('gemini_cli_quota.missing_project_id'));
-  }
-
-  const quotaResponse = await apiCallApi.request({
-    authIndex,
-    method: 'POST',
-    url: GEMINI_CLI_QUOTA_URL,
-    header: { ...GEMINI_CLI_REQUEST_HEADERS },
-    data: JSON.stringify({ project: projectId }),
-  });
-  if (quotaResponse.statusCode < 200 || quotaResponse.statusCode >= 300) {
-    throw createStatusError(getApiCallErrorMessage(quotaResponse), quotaResponse.statusCode);
-  }
-
-  const payload = parseGeminiCliQuotaPayload(quotaResponse.body ?? quotaResponse.bodyText);
-  const buckets = Array.isArray(payload?.buckets) ? payload?.buckets : [];
-
-  const parsedBuckets = buckets
-    .map((bucket) => {
-      const modelId = normalizeGeminiCliModelId(bucket.modelId ?? bucket.model_id);
-      if (!modelId) return null;
-      const tokenType = normalizeStringValue(bucket.tokenType ?? bucket.token_type);
-      const remainingFractionRaw = normalizeQuotaFraction(
-        bucket.remainingFraction ?? bucket.remaining_fraction
-      );
-      const remainingAmount = normalizeNumberValue(
-        bucket.remainingAmount ?? bucket.remaining_amount
-      );
-      const resetTime = normalizeStringValue(bucket.resetTime ?? bucket.reset_time) ?? undefined;
-      let fallbackFraction: number | null = null;
-      if (remainingAmount !== null) {
-        fallbackFraction = remainingAmount <= 0 ? 0 : null;
-      } else if (resetTime) {
-        fallbackFraction = 0;
-      }
-      const remainingFraction = remainingFractionRaw ?? fallbackFraction;
-      return {
-        modelId,
-        tokenType,
-        remainingFraction,
-        remainingAmount,
-        resetTime,
-      };
-    })
-    .filter((bucket): bucket is GeminiCliParsedBucket => bucket !== null);
-
-  const builtBuckets = buildGeminiCliQuotaBuckets(parsedBuckets);
+  const quotaData = await fetchGeminiCliQuotaBuckets(file, t);
   const supplementaryRequestId = scheduleGeminiCliSupplementaryRefresh(
     file.name,
-    authIndex,
-    projectId,
+    quotaData.authIndex,
+    quotaData.projectId,
     t
   );
   const supplementarySnapshot = readGeminiCliSupplementarySnapshot(
@@ -510,7 +184,7 @@ const fetchGeminiCliQuota = async (
   return {
     fileName: file.name,
     supplementaryRequestId,
-    buckets: builtBuckets,
+    buckets: quotaData.buckets,
     tierLabel: supplementarySnapshot.tierLabel,
     tierId: supplementarySnapshot.tierId,
     creditBalance: supplementarySnapshot.creditBalance,
@@ -766,141 +440,6 @@ const renderGeminiCliItems = (
   return h(Fragment, null, ...nodes);
 };
 
-const buildClaudeQuotaWindows = (
-  payload: ClaudeUsagePayload,
-  t: TFunction
-): ClaudeQuotaWindow[] => {
-  const windows: ClaudeQuotaWindow[] = [];
-
-  for (const { key, id, labelKey } of CLAUDE_USAGE_WINDOW_KEYS) {
-    const window = payload[key as keyof ClaudeUsagePayload];
-    if (!window || typeof window !== 'object' || !('utilization' in window)) continue;
-    const typedWindow = window as { utilization: number; resets_at: string };
-    const usedPercent = normalizeNumberValue(typedWindow.utilization);
-    const resetLabel = formatQuotaResetTime(typedWindow.resets_at);
-    windows.push({
-      id,
-      label: t(labelKey),
-      labelKey,
-      usedPercent,
-      resetLabel,
-    });
-  }
-
-  return windows;
-};
-
-const normalizeFlagValue = (value: unknown): boolean | undefined => {
-  if (value === undefined || value === null) return undefined;
-  if (typeof value === 'boolean') return value;
-  if (typeof value === 'number') return value !== 0;
-  if (typeof value === 'string') {
-    const trimmed = value.trim().toLowerCase();
-    if (['true', '1', 'yes', 'y', 'on'].includes(trimmed)) return true;
-    if (['false', '0', 'no', 'n', 'off'].includes(trimmed)) return false;
-  }
-  return undefined;
-};
-
-const parseClaudeProfilePayload = (payload: unknown): ClaudeProfileResponse | null => {
-  if (payload === undefined || payload === null) return null;
-  if (typeof payload === 'string') {
-    const trimmed = payload.trim();
-    if (!trimmed) return null;
-    try {
-      return JSON.parse(trimmed) as ClaudeProfileResponse;
-    } catch {
-      return null;
-    }
-  }
-  if (typeof payload === 'object') {
-    return payload as ClaudeProfileResponse;
-  }
-  return null;
-};
-
-const resolveClaudePlanType = (profile: ClaudeProfileResponse | null): string | null => {
-  if (!profile) return null;
-
-  const hasClaudeMax = normalizeFlagValue(profile.account?.has_claude_max);
-  if (hasClaudeMax) return 'plan_max';
-
-  const hasClaudePro = normalizeFlagValue(profile.account?.has_claude_pro);
-  if (hasClaudePro) return 'plan_pro';
-
-  const organizationType = normalizeStringValue(
-    profile.organization?.organization_type
-  )?.toLowerCase();
-  const subscriptionStatus = normalizeStringValue(
-    profile.organization?.subscription_status
-  )?.toLowerCase();
-
-  if (organizationType === 'claude_team' && subscriptionStatus === 'active') {
-    return 'plan_team';
-  }
-
-  if (hasClaudeMax === false && hasClaudePro === false) return 'plan_free';
-
-  return null;
-};
-
-const fetchClaudeQuota = async (
-  file: AuthFileItem,
-  t: TFunction
-): Promise<{
-  windows: ClaudeQuotaWindow[];
-  extraUsage?: ClaudeExtraUsage | null;
-  planType?: string | null;
-}> => {
-  const rawAuthIndex = file['auth_index'] ?? file.authIndex;
-  const authIndex = normalizeAuthIndex(rawAuthIndex);
-  if (!authIndex) {
-    throw new Error(t('claude_quota.missing_auth_index'));
-  }
-
-  const [usageResult, profileResult] = await Promise.allSettled([
-    apiCallApi.request({
-      authIndex,
-      method: 'GET',
-      url: CLAUDE_USAGE_URL,
-      header: { ...CLAUDE_REQUEST_HEADERS },
-    }),
-    apiCallApi.request({
-      authIndex,
-      method: 'GET',
-      url: CLAUDE_PROFILE_URL,
-      header: { ...CLAUDE_REQUEST_HEADERS },
-    }),
-  ]);
-
-  if (usageResult.status === 'rejected') {
-    throw usageResult.reason;
-  }
-
-  const result = usageResult.value;
-
-  if (result.statusCode < 200 || result.statusCode >= 300) {
-    throw createStatusError(getApiCallErrorMessage(result), result.statusCode);
-  }
-
-  const payload = parseClaudeUsagePayload(result.body ?? result.bodyText);
-  if (!payload) {
-    throw new Error(t('claude_quota.empty_windows'));
-  }
-
-  const windows = buildClaudeQuotaWindows(payload, t);
-  const planType =
-    profileResult.status === 'fulfilled' &&
-    profileResult.value.statusCode >= 200 &&
-    profileResult.value.statusCode < 300
-      ? resolveClaudePlanType(
-          parseClaudeProfilePayload(profileResult.value.body ?? profileResult.value.bodyText)
-        )
-      : null;
-
-  return { windows, extraUsage: payload.extra_usage, planType };
-};
-
 const renderClaudeItems = (
   quota: ClaudeQuotaState,
   t: TFunction,
@@ -1114,32 +653,6 @@ export const GEMINI_CLI_CONFIG: QuotaConfig<
   controlClassName: styles.geminiCliControl,
   gridClassName: styles.geminiCliGrid,
   renderQuotaItems: renderGeminiCliItems,
-};
-
-const fetchKimiQuota = async (file: AuthFileItem, t: TFunction): Promise<KimiQuotaRow[]> => {
-  const rawAuthIndex = file['auth_index'] ?? file.authIndex;
-  const authIndex = normalizeAuthIndex(rawAuthIndex);
-  if (!authIndex) {
-    throw new Error(t('kimi_quota.missing_auth_index'));
-  }
-
-  const result = await apiCallApi.request({
-    authIndex,
-    method: 'GET',
-    url: KIMI_USAGE_URL,
-    header: { ...KIMI_REQUEST_HEADERS },
-  });
-
-  if (result.statusCode < 200 || result.statusCode >= 300) {
-    throw createStatusError(getApiCallErrorMessage(result), result.statusCode);
-  }
-
-  const payload = parseKimiUsagePayload(result.body ?? result.bodyText);
-  if (!payload) {
-    throw new Error(t('kimi_quota.empty_data'));
-  }
-
-  return buildKimiQuotaRows(payload);
 };
 
 const renderKimiItems = (
